Script to generate images for runs
This is adapted from a gist authored by Dan Allan available at
https://gist.github.com/danielballan/c46a356067b4ad423bc5de6f183cf474
with some tweaks to rescale to the full 8 bit unsigned range. To use it
with the server iterate over all runs and place the PNGs in /tmp/, there
are many ways that this can and should be improved but it demonstrates
the concept well in a running instance.