Return a richer set of run data
I added this a few weeks ago, and forgot to push it. It is backwards
compatible with the old returned data as it just adds more. This is
intended to support a richer result page when looking at runs in a
catalog, and eventually when searching runs in a catalog.